requirements_colab.txt
annoy==1.17.0
boto3==1.17.85
python-dotenv==0.17.1
pytorch-lightning==1.3.3
wandb==0.10.31
